Update README.md
Copying the content from alluxio.org created some broken links and didn't describe the purpose of this repository. It also makes it difficult to keep the content here in sync with the content on the website. It's better for this repo to have its own description which just references the original content on alluxio.org.